moved tabs to a new component
