+++ src/components/Admin/Admin.jsx
@@ -1,29 +1,14 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './Admin.css';
 
 import AdminHeader from './AdminHeader/AdminHeader';
 import AdminCreateButton from './AdminCreateButton/AdminCreateButton';
 import AdminLogOutButton from './AdminLogOutButton/AdminLogOutButton';
-// import AdminArtistTable from './AdminArtistTable/AdminArtistTable';
-
-// material-ui imports
-import { makeStyles } from '@material-ui/core/styles';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import { AppBar } from '@material-ui/core';
-import AdminArtistTable from './AdminArtistTable/AdminArtistTable';
-
-const useStyles = makeStyles({
-    root: {
-        flexGrow: 1,
-    },
-});
+import AdminTabs from './AdminTabs/AdminTabs';
 
 // function for Admin component
 function Admin() {
 
-    // set material-ui classes variable
-    const classes = useStyles();
 
     return (
         <div>
@@ -41,19 +26,10 @@ function Admin() {
             <div>
                 <h2>Admin</h2>
             </div>
-            <div className={classes.root}>
-                <AppBar position="static" color="default">
-                    <Tabs
-                        indicatorColor="primary"
-                        textColor="primary"
-                        centered
-                    >
-                        <Tab label="Orders" />
-                        <Tab label="Issues" />
-                        <Tab label="Artists" />
-                    </Tabs>
-                </AppBar>
+            <div>
+                <AdminTabs />
             </div>
+            
         </div>
     )
 }
